Moving loadcell reading to separate file, sending entire partition in characteristic.
