+++ partitions.csv
@@ -5,4 +5,4 @@ phy_init, data, phy,     ,        0x1000,
 factory,  app,  factory, ,        1M,
 ota_0,    app,  ota_0,   ,        1M,
 ota_1,    app,  ota_1,   ,        1M,
-loadcell_calib,   0x40, 0x06,    ,        0x1000, readonly
+loadcell_calib,   0x40, 0x06,    ,        0xff, readonly
